Remove methods related to "original" values from `EntityInterface`.
These methods have limited usage and "original" value tracking can be considered an
optional feature of the entity. The removed methods have been moved into the new
`OriginalValueInterface`.
