+++ src/ORM/Entity.php
@@ -19,12 +19,13 @@ namespace Cake\ORM;
 use Cake\Datasource\EntityInterface;
 use Cake\Datasource\EntityTrait;
 use Cake\Datasource\InvalidPropertyInterface;
+use Cake\Datasource\OriginalValueInterface;
 
 /**
  * An entity represents a single result row from a repository. It exposes the
  * methods for retrieving and storing properties associated in this row.
  */
-class Entity implements EntityInterface, InvalidPropertyInterface
+class Entity implements EntityInterface, InvalidPropertyInterface, OriginalValueInterface
 {
     use EntityTrait;
